feat: refactor SQL recording state management
